+++ eras/babbage/impl/src/Cardano/Ledger/Babbage/TxInfo.hs
@@ -54,7 +54,6 @@ import Cardano.Ledger.Babbage.Scripts (PlutusScript (..))
 import Cardano.Ledger.Babbage.UTxO ()
 import Cardano.Ledger.BaseTypes (
   Inject (..),
-  ProtVer,
   StrictMaybe (..),
   isSJust,
   kindObjectValue,
@@ -193,21 +192,21 @@ transRedeemer :: Data era -> PV2.Redeemer
 transRedeemer = PV2.Redeemer . PV2.dataToBuiltinData . getPlutusData
 
 transRedeemerPointerV2V3 ::
-  forall proxy l era t.
-  ( EraPlutusTxInfo l era
+  forall proxy l era.
+  ( EraTx era
   , AlonzoEraTxBody era
+  , EraPlutusTxInfo l era
   , Inject (BabbageContextError era) (ContextError era)
   ) =>
   proxy l ->
-  ProtVer ->
-  TxBody t era ->
+  LedgerTxInfo era ->
   (PlutusPurpose AsIx era, (Data era, ExUnits)) ->
   Either (ContextError era) (PlutusScriptPurpose l, PV2.Redeemer)
-transRedeemerPointerV2V3 proxy pv txBody (ptr, (d, _)) =
-  case redeemerPointerInverse txBody ptr of
+transRedeemerPointerV2V3 proxy lti@LedgerTxInfo {ltiTx} (ptr, (d, _)) =
+  case redeemerPointerInverse (ltiTx ^. bodyTxL) ptr of
     SNothing -> Left $ inject $ RedeemerPointerPointsToNothing ptr
     SJust sp -> do
-      plutusScriptPurpose <- toPlutusScriptPurpose proxy pv sp
+      plutusScriptPurpose <- toPlutusScriptPurpose proxy lti sp
       Right (plutusScriptPurpose, transRedeemer d)
 
 -- | Translate all `Redeemers` from within a `Tx` into a Map from a `PlutusScriptPurpose`
@@ -220,14 +219,13 @@ transTxRedeemers ::
   , Inject (BabbageContextError era) (ContextError era)
   ) =>
   proxy l ->
-  ProtVer ->
-  Tx t era ->
+  LedgerTxInfo era ->
   Either (ContextError era) (PV2.Map (PlutusScriptPurpose l) PV2.Redeemer)
-transTxRedeemers proxy pv tx =
+transTxRedeemers proxy lti@LedgerTxInfo {ltiTx} =
   PV2.unsafeFromList
     <$> mapM
-      (transRedeemerPointerV2V3 proxy pv (tx ^. bodyTxL))
-      (Map.toList $ tx ^. witsTxL . rdmrsTxWitsL . unRedeemersL)
+      (transRedeemerPointerV2V3 proxy lti)
+      (Map.toList $ ltiTx ^. witsTxL . rdmrsTxWitsL . unRedeemersL)
 
 instance EraPlutusContext BabbageEra where
   type ContextError BabbageEra = BabbageContextError BabbageEra
@@ -327,7 +325,7 @@ instance ToJSON (PlutusPurpose AsIx era) => ToJSON (BabbageContextError era) whe
 instance EraPlutusTxInfo 'PlutusV1 BabbageEra where
   toPlutusTxCert _ _ = pure . Alonzo.transTxCert
 
-  toPlutusScriptPurpose = Alonzo.transPlutusPurpose
+  toPlutusScriptPurpose proxy lti = Alonzo.transPlutusPurpose proxy (ltiProtVer lti)
 
   toPlutusTxInfo proxy LedgerTxInfo {ltiProtVer, ltiEpochInfo, ltiSystemStart, ltiUTxO, ltiTx} =
     PlutusTxInfoResult $ withTopTxLevelOnly ltiTx $ \tx -> do
@@ -369,9 +367,9 @@ instance EraPlutusTxInfo 'PlutusV1 BabbageEra where
 instance EraPlutusTxInfo 'PlutusV2 BabbageEra where
   toPlutusTxCert _ _ = pure . Alonzo.transTxCert
 
-  toPlutusScriptPurpose = Alonzo.transPlutusPurpose
+  toPlutusScriptPurpose proxy lti = Alonzo.transPlutusPurpose proxy (ltiProtVer lti)
 
-  toPlutusTxInfo proxy LedgerTxInfo {ltiProtVer, ltiEpochInfo, ltiSystemStart, ltiUTxO, ltiTx} =
+  toPlutusTxInfo proxy lti@LedgerTxInfo {ltiProtVer, ltiEpochInfo, ltiSystemStart, ltiUTxO, ltiTx} =
     PlutusTxInfoResult $ withTopTxLevelOnly ltiTx $ \tx -> do
       let txBody = tx ^. bodyTxL
       timeRange <-
@@ -384,7 +382,7 @@ instance EraPlutusTxInfo 'PlutusV2 BabbageEra where
           [minBound ..]
           (F.toList (txBody ^. outputsTxBodyL))
       txCerts <- Alonzo.transTxBodyCerts proxy ltiProtVer txBody
-      plutusRedeemers <- transTxRedeemers proxy ltiProtVer tx
+      plutusRedeemers <- transTxRedeemers proxy lti
       -- It is important for memoization for `txInfo` to be a let binding
       let
         txInfo =
@@ -418,11 +416,11 @@ toPlutusV2Args ::
   PlutusPurpose AsIxItem era ->
   Data era ->
   Either (ContextError era) (PlutusArgs 'PlutusV2)
-toPlutusV2Args proxy LedgerTxInfo {..} txInfo scriptPurpose redeemerData =
+toPlutusV2Args proxy lti@LedgerTxInfo {..} txInfo scriptPurpose redeemerData =
   PlutusV2Args
     <$> toLegacyPlutusArgs
       proxy
-      ltiProtVer
+      lti
       (PV2.ScriptContext txInfo)
       scriptPurpose
       maybeSpendingDatum